Add files via upload
The aim of this project is to build a machine learning model to predict whether or not a NASA classified NEO (Nearest Earth Object) is hazardous or not. My data is taken from Kaggle where it was compiled from the NASA open API and the Jet Propulsion Lab Center for Near Earth Object Studies. Throughout this project, we will implement multiple machine learning techniques to yield the best results.